Use `grind =>` instead of `grind .` for allocEmpty (#68)
Using `grind =>` means that `h -> x = y` will be used by grind if `h`
and `x` are in the context. Previously, with `grind .`, this lemma was
added in the context if both `h`, `x`, and `y` were in the context.

For `allocEmpty`, this means that upon seeing `allocEmpty ctx = some
(ctx', op')` and `ptr.get! ctx'`, the value of `ptr.get! ctx` will be
added in the grind context. Previously, it was required for `ptr.get!
ctx` to appear to be added in the context, which was preventing some
`grind` to succeed in future PRs.